+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Bug #39: Resolve `Container::get()` type inference for unconfigured classes in config (`ServiceMap`) (@terabytesoftw)
 - Enh #40: Enhance `PHPStan` analysis for `Behavior` type inference and testing (@terabytesoftw)
 - Enh #41: Refactor `PHPDoc` comments for consistency and clarity (@terabytesoftw)
+- Bug #42: Move `ApplicationPropertiesClassReflectionExtension` to `property` directory and add testing (@terabytesoftw)
 
 ## 0.2.3 June 09, 2025